Move ssr vm-modules demo into its own dir
