+++ packages/labs/ssr/package.json
@@ -10,7 +10,7 @@
   "scripts": {
     "build": "tsc",
     "build:watch": "tsc --watch",
-    "start": "node --experimental-vm-modules ./demo/server.js",
+    "demo:vm-modules": "node --experimental-vm-modules ./demo/vm-modules/server.js",
     "test": "npm run test:unit && npm run test:integration",
     "test:integration": "npm run test:integration:dev && npm run test:integration:prod",
     "test:integration:dev": "MODE=dev NODE_OPTIONS=--experimental-vm-modules wtr",
