+++ packages/labs/ssr/src/demo/vm-modules/app-client.ts
