+++ packages/labs/ssr/src/demo/vm-modules/module.ts
